Sale refactor

+++ app/Http/Controllers/VoucherController.php
@@ -10,20 +10,21 @@ use App\Models\VoucherRecord;
 use Carbon\Carbon;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
-class VoucherController extends Controller
-{
+class VoucherController extends Controller {
     /**
      * Display a listing of the resource.
      */
     public function index()
     {
-        if (Auth::user()->role == "admin") {
+        if (Gate::allows('isAdmin')) {
             $vouchers = Voucher::latest("id")
-                ->paginate(7)->withQueryString();
+                ->paginate(15)->withQueryString();
         } else {
             // $vouchers = Auth::user()->vouchers()->whereDate('created_at', Carbon::today())->get();
-            $vouchers = Voucher::where("user_id", Auth::id())->whereDate('created_at', Carbon::today())->get();
+            // $vouchers = Voucher::where("user_id", Auth::id())->whereDate('created_at', Carbon::today())->get();
+            $vouchers = Auth::user()->vouchers()->whereDate('created_at', Carbon::today())->get();
         }
         return VoucherResource::collection($vouchers);
         //
@@ -36,8 +37,10 @@ class VoucherController extends Controller
     {
         // return $request;
         $request->validate([
-            'customer_name' => 'required|min:5|max:20',
-            'phone_number' => 'required',
+            // @fix makes customer_name optional
+            'customer_name' => 'nullable',
+            'phone_number' => 'nullable',
+            'voucher_number' => 'required',
             'records' => 'array',
             'records.*.product_id' => 'required|exists:products,id',
             'records.*.quantity' => 'required|numeric|min:1'
@@ -47,14 +50,18 @@ class VoucherController extends Controller
         $voucher = new Voucher;
         $voucher->customer_name = $request->customer_name;
         $voucher->phone_number = $request->phone_number;
+        $voucher->voucher_number = $request->voucher_number;
         $voucher->user_id = Auth::id();
         $voucher->save();
 
         // creating voucher records
+        // @fix, use Product::inWhere to optimize query
+
         $total = 0;
         foreach ($request->records as $record) {
             $product = Product::find($record['product_id']);
             $quantity = $record['quantity'];
+            // @fix, return error/informative message on insufficient stock
             if ($product->total_stock >= $quantity) {
                 $cost = $product->sale_price * $quantity;
                 VoucherRecord::create([
@@ -69,7 +76,7 @@ class VoucherController extends Controller
 
         // update voucher
         $voucher->total = $total;
-        $voucher->tax = $total * 0.2;
+        $voucher->tax = $total * 0.02;
         $voucher->net_total = $total + $voucher->tax;
         $voucher->save();
 
@@ -83,34 +90,6 @@ class VoucherController extends Controller
      */
     public function show(string $id)
     {
-
-        $voucher = Voucher::find($id);
-
-        if (is_null($voucher)) {
-            return response()->json([
-                // "success" => false,
-                "message" => "Product not found",
-
-            ], 404);
-        }
-        return new VoucherDetailResource($voucher);
-        //
-    }
-
-    /**
-     * Update the specified resource in storage.
-     */
-    public function update(Request $request, string $id)
-    {
-        //
-    }
-
-    /**
-     * Remove the specified resource from storage.
-     */
-    public function destroy(string $id)
-    {
-
         $voucher = Voucher::find($id);
         if (is_null($voucher)) {
             return response()->json([
@@ -120,27 +99,96 @@ class VoucherController extends Controller
             ], 404);
         }
 
+        $this->authorize('view', $voucher);
+        return new VoucherDetailResource($voucher);
+        //
+    }
+
+    /**
+     * Remove the specified resource from storage.
+     */
+    public function destroy(string $id)
+    {
+        $voucher = Voucher::find($id);
+        if (is_null($voucher)) {
+            return response()->json([
+                // "success" => false,
+                "message" => "Voucher not found",
+
+            ], 404);
+        }
+
+        $this->authorize('delete', $voucher);
         $voucher->delete();
 
         return response()->json([
             "message" => "you have deleted voucher"
-        ]);
+        ], 204);
     }
 
 
     public function restore($id)
     {
-        $softDeletedVoucher = Voucher::withTrashed()->find($id);
-        if (is_null($softDeletedVoucher)) {
+        $voucher = Voucher::onlyTrashed()->find($id);
+        if (is_null($voucher)) {
             return response()->json([
-                'message' => 'SoftdeletedVoucher is not found'
+                'message' => 'Voucher does on exist'
             ]);
         }
 
-        $softDeletedVoucher->restore();
+        $this->authorize('restore', $voucher);
+        $voucher->restore();
 
         return response()->json([
-            'message' => "softdeletedVoucher has been restored"
-        ]);
+            'message' => "Voucher has been restored"
+        ], 201);
+    }
+
+    public function showTrash()
+    {
+        if (Gate::allows('isAdmin')) {
+            $trashed_vouchers = Voucher::onlyTrashed()->get();
+        } else {
+            $trashed_vouchers = Auth::user()->vouchers()->onlyTrashed()->get();
+        }
+
+        return VoucherResource::collection($trashed_vouchers);
+    }
+
+    // isAdmin middleware
+    public function forceDelete(string $id)
+    {
+        $voucher = Voucher::onlyTrashed()->find($id);
+        if (is_null($voucher)) {
+            return response()->json([
+                'message' => 'Voucher does on exist'
+            ]);
+        }
+
+        // @fix
+        $voucher->forceDelete();
+        return response()->json([
+            'message' => 'You have deleted voucher permanently'
+        ], 204);
+    }
+
+    // isAdmin middleware
+    public function emptyBin()
+    {
+        Voucher::onlyTrashed()->forceDelete();
+        return response()->json([
+            'message' => 'Trash has been emptied'
+        ], 204);
+
+    }
+
+    // isAdmin middleware
+    public function recycleBin()
+    {
+        Voucher::onlyTrashed()->restore();
+
+        return response()->json([
+            'message' => 'Trash has been restored'
+        ], 201);
     }
 }
